requirements.txt
openai-whisper==20240930
groq~=0.11.0
python-dotenv==1.0.1
python-espeak